requirements.txt
# Environment: myvenv
langchain==1.2.0
python-dotenv==1.2.1
ipykernel==7.1.0
langchain-groq==1.1.1
langchain-google-genai==4.1.1
langchain-community==0.4.1
langchain-text-splitters==1.1.0
pypdf==6.4.2
faiss-cpu==1.13.1
pytest==9.0.2
PyMuPDF==1.26.7
black==25.12.0
isort==7.0.0
flake8==7.3.0
ai-foundation-kit==0.1.6